Move ctaType to shared utils
